Extend from abstract-stream-down for easier query support
This will also allow us to easily support the various batch modes including
createWriteStream (which would be ideal w/ the way things are set up).
